third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-evaluate-in-default-context.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

function test()
{
    InspectorTest.sendCommand("Runtime.evaluate", { "expression": "window.a" }, evaluateCallback);

    function evaluateCallback(result)
    {
        result.id = 0;
        InspectorTest.logObject(result);
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body>
Tests that default execution context is selected correctly.
<iframe src="resources/runtime-events-iframe.html" onLoad="runTest();"></iframe>
</body>
</html>
