third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-evaluate-in-default-context-expected.txt

Tests that default execution context is selected correctly. 
{
    id : 0
    result : {
        result : {
            description : 42
            type : number
            value : 42
        }
    }
}
